pyproject.toml
[project]
name = "sc_fa"
version = "0.1.5"
description = "HTTP URI short-cutter API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocsv>=1.4.0",
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.11.0",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
    "typer>=0.21.1",
    "types-aiofiles>=25.1.0.20251011",
]

[build-system]
requires = ["hatchling>=1.11.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["app/*", "cli.py"]

[project.scripts]
sc-cli = "app.cli:app"
